Redesign of CONTENT vs FORMAT
Now a FORMAT is more aligned with what formats really are and the same applies to CONTENT. Internal format introduced that represents a serialization of the internal table used to store and transfer data. Introduction of a more powerful and generic tool for parsing parameters. Several other changes and improvements.